Issue-2245: Handle proxy fields in API field updates (#2246) + JSONAPI
* Handle proxy fields in API field updates

* PEP8

* PEP8

* use api for schema lookup

* unpinned zc.buildout to make the tests pass

* get schema by the usual way to fix the test

* Check if we have a proxy object

Conflicts:
	CHANGES.rst
	bika/lims/jsonapi/update.py
	bika/lims/jsonapi/v1/__init__.py
	buildout.cfg
	travis.cfg
